Added support for centos and support for python 2.6.6

+++ scripts/installessentials.sh
@@ -1,5 +1,16 @@
 # Installing required packages
-apt-get install python-dev libffi-dev libssl-dev libxml2-dev libxslt1-dev libjpeg8-dev zlib1g-dev make cmake python-setuptools gcc -y
+
+if [ -f "/usr/bin/apt-get" ]
+then
+  apt-get install python-dev libffi-dev libssl-dev libxml2-dev libxslt1-dev libjpeg8-dev zlib1g-dev make cmake python-setuptools gcc -y
+elif [ -f "/usr/bin/yum" ]
+then
+  yum install python-devel libffi-devel openssl-devel libxml2-devel libxslt-devel libjpeg-devel zlib-devel make cmake python-setuptools gcc -y
+else
+  echo "No package managers found ..."
+  exit
+fi
+
 easy_install pip
 
 # Creating required directories
